int to size_t

+++ vigenere.c
@@ -95,8 +95,8 @@ void decrypt(const char *passPhrase, const char *encryptedFileName) {
  */
 void checkRepetitionInString(char *string) {
   size_t len = strlen(string);
-  int i = 0;
-  int x = 1;
+  size_t i = 0;
+  size_t x = 1;
   while (i + x < len) {
     if (string[i] == string[i + x]) {
       i += x;
